requirements.txt
numpy
langchain
langchain-groq
langchain_core
langchain_community
pymongo
python-dotenv